Move the `MemberList` into the view
